agregando validaciones
